Started parsing form + typo
Started writing the function to parse the form data.
To log the first three field's value to the console, press the 'pulzante'.
I'm not satisfacted with the way I'm accessing each field value
I.e. by index and not by a key-string, like the field name. I've to investigate further.

Fixed a small typo in dev_note.md